dump log if integration test fails

+++ scripts/travis/build-all-in-one-image.sh
@@ -13,6 +13,14 @@ export REPO=jaegertracing/all-in-one
 docker build -f cmd/all-in-one/Dockerfile -t $REPO:latest cmd/all-in-one
 export CID=$(docker run -d -p 16686:16686 -p 5778:5778 $REPO:latest)
 make integration-test
+
+if [ -z "$( docker ps | grep jaeger )" ] ; then
+    echo "---- jaeger stopped unexpectedly during test ---------"
+    echo "---- please check the following log for details ----"
+    docker logs $CID
+    exit 1
+fi
+
 docker kill $CID
 
 # Only push the docker container to Docker Hub for master branch
